requirements.txt
llama-index==0.11.23
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-embeddings-ollama
llama-index-llms-ollama
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pandas==2.2.2
